bcachefs: split out lru_format.h
